// PIM/Employment-List/Personal-Details/Personal-Details-Type.ts
export type PersonalDetailsType = {
    firstName: string;
    middleName: string;
    lastName: string;
    employeeId: string;
    otherId: string;
    licenseNumber: string;
    licenseExpiryDate: string; 
    nationality: string;
    maritalStatus: string;
    dateOfBirth: string; 
    Gender: 'Male' | 'Female';
}
    